D3D11

+++ code/Common/md5.hxx
@@ -137,7 +137,7 @@ void MD5::update(ifstream& in) {
 		in.read(buffer, BUFFER_SIZE);
 		length = in.gcount();
 		if (length > 0) {
-			update(buffer, length);
+			update(buffer, (size_t)length);
 		}
 	}
 	in.close();
